==> 1. Data collection.ipynb <==
# This notebook integrates downloaded data from website www.360che.com

import pandas as pd
import numpy as np

import json
from pathlib import Path

def main():
    for q in ['平台', '满帮', '货拉拉', '货车帮', '运满满']:
        for p in range(1, 101):
            fpath = Path(f'Data/down/index/{q}.{p}.json')
            if not fpath.exists():
                break
            index = json.loads(fpath.read_text())
            posts = index['data']
            if not posts:
                break
            print(posts[0].keys())
            for post in posts:
                print(post['title'], post['date'], post['score'])
main()


fpath = Path('Data/down/index/平台.1.json')
index = json.loads(fpath.read_text())

pd.DataFrame(index['data'])

fpath = Path('Data/down/index/平台.1.json')
index = json.loads(fpath.read_text())
pd.DataFrame(index['data'])

path = Path('Data/down/content-parsed.json')
content = json.loads(path.read_text())

index = []
keys_to_keep = ['tid', 'author', 'description', 
                'image', 'replies', 'timestamp', 'title', 'views', 
                'date', 'thread_url', 'pc_thread_url', 'nickname', 'circleId', 'circleName', 'threadType', 'topicId', 
                'topicName', 'images', 'isHot', 'score']
for q in ['平台', '满帮', '货拉拉', '货车帮', '运满满']:
    for p in range(1, 101):
        fpath = Path(f'Data/down/index/{q}.{p}.json')
        if not fpath.exists():
            break
        data = json.loads(fpath.read_text())
        assert data['msg'] == '成功'
        data = [{key: i[key] for key in keys_to_keep} for i in data['data']]
        # [f(i) for i in x] ==> result = []; for i in x: result.append(f(i))
        for i in data:
            i['keyword'] = q
            if i['pc_thread_url'] in content:
                i['description'] = content[i['pc_thread_url']]
        index.extend(data)
df = pd.DataFrame(index)

df['threadType'].value_counts()

df['author'].value_counts()

df.to_parquet('RawData.parquet')
df.to_excel('RawData.xlsx')
