==> 4. EDA.ipynb <==
# To-do list:
1. Use tf-idf to select Top words: Top 30 words for each platform (yinxian zhang's paper)
2. Depict changes of top words over time
3. Word2vec--what words people use when talking about 'pingtai' (yinxian zhang's paper)
4. [Done] 一个假设：跑平台的司机倾向于某种车型，比如轻卡，因为成本低。“重卡”关键词已被dropped
5. 疫情相关内容
6. [Done] Manual coding for all tasks and calculate the accuracy rate of ChatGPT


# Data Cleaning

import pandas as pd
import numpy as np

df = pd.read_parquet('../processed_data/tokenized_data.parquet')
df.info()

# df.rename(columns ={'tok':'hanlp_tok'}, inplace=True)

df['clean_tok']

for i in ['clean_tok']:
    df[i] = df[i].apply(lambda f: f.tolist())

# Clean timestamp--separate year and month
df['date'].head()

df['year'] = df['date'].apply(lambda x: x[:4])

df['year'] = pd.to_numeric(df['year'])

import seaborn as sns
sns.histplot(df['year'])

df.to_parquet('../processed_data/tokenized_data.parquet')

# Top-frequency words

# Word counts
words = sum(df['clean_tok'], [])
from collections import Counter
ct = Counter(words)
print(ct.most_common(20))

# word count by keywords
for keyword in df['keyword'].unique():
    df_keyword = df[df['keyword'] == keyword]
    words_keyword = sum(df_keyword['clean_tok'], [])
    ct_keyword = Counter(words_keyword)
    print(f"Top 30 words for {keyword}:")
    print(ct_keyword.most_common(30))
    print("\n") 

# word count by year
for year in df['year'].unique():
    df_year = df[df['year'] == year]
    words_year = sum(df_year['clean_tok'], [])
    ct_year = Counter(words_year)
    print(f"Top 30 words for {year}:")
    print(ct_year.most_common(20))
    print("\n")
